Camera value updates & mapview implem. changes
I Split the existing mapView extension in MapViewExtension and FlutterMapView which extends MKMapView.
FlutterMapView contains overrides which do not belong in an Extension. MapViewExtension contains the
zoomLevel implementation. Coordinate updates where corrected. Heading values are now constantly calculated.
There is currently no way to get an exact and continuous update of the tilt and I have to still rely on the
standard mapView camera.